requirements.txt
pandas==1.5.3
numpy==1.23.5
faiss-cpu==1.7.3
sentence-transformers==2.2.2